+++ src/hasSameValues.ts
@@ -1,15 +1,36 @@
+import isSync from "./_isSync.ts";
 import quickSize from "./_quickSize.ts";
 
-const countValues = <T>(iterable: Iterable<T>): Map<T, number> => {
-  const result = new Map<T, number>();
-  for (const value of iterable) {
-    result.set(value, (result.get(value) || 0) + 1);
-  }
-  return result;
-};
+export default hasSameValues;
+
+/** @ignored */
+function hasSameValues<T>(
+  iterableA: Iterable<T>,
+  iterableB: Iterable<T>,
+): boolean;
+
+/** @ignored */
+function hasSameValues<T>(
+  iterableA: AsyncIterable<T>,
+  iterableB: Iterable<T>,
+): Promise<boolean>;
+
+/** @ignored */
+function hasSameValues<T>(
+  iterableA: Iterable<T>,
+  iterableB: AsyncIterable<T>,
+): Promise<boolean>;
+
+/** @ignored */
+function hasSameValues<T>(
+  iterableA: AsyncIterable<T>,
+  iterableB: AsyncIterable<T>,
+): Promise<boolean>;
 
 /**
  * If `iterableA` and `iterableB` have the same lengths and values, returns `true` (order is irrelevant). Otherwise, returns `false`. Equality is determined with `Object.is`.
+ *
+ * Works with sync and async iterables. If passed an async iterable, returns a Promise for the result.
  *
  * @example
  * ```typescript
@@ -28,7 +49,48 @@ const countValues = <T>(iterable: Iterable<T>): Map<T, number> => {
  * // => true
  * ```
  */
-export default function hasSameValues<T>(
+function hasSameValues<T>(
+  iterableA: Iterable<T> | AsyncIterable<T>,
+  iterableB: Iterable<T> | AsyncIterable<T>,
+): boolean | Promise<boolean> {
+  return (
+      isSync(iterableA) && isSync(iterableB)
+    )
+    ? (
+      hasSameValuesSync(iterableA, iterableB)
+    )
+    : (
+      hasSameValuesAsync(iterableA, iterableB)
+    );
+}
+
+function countValuesSync<T>(iterable: Iterable<T>): Map<T, number> {
+  const result = new Map<T, number>();
+  for (const value of iterable) {
+    result.set(value, (result.get(value) || 0) + 1);
+  }
+  return result;
+}
+
+async function countValuesAsync<T>(
+  iterable: AsyncIterable<T>,
+): Promise<Map<T, number>> {
+  const result = new Map<T, number>();
+  for await (const value of iterable) {
+    result.set(value, (result.get(value) || 0) + 1);
+  }
+  return result;
+}
+
+function countValuesEither<T>(
+  iterable: Iterable<T> | AsyncIterable<T>,
+): Map<T, number> | Promise<Map<T, number>> {
+  return isSync(iterable)
+    ? countValuesSync(iterable)
+    : countValuesAsync(iterable);
+}
+
+function hasSameValuesSync<T>(
   iterableA: Iterable<T>,
   iterableB: Iterable<T>,
 ): boolean {
@@ -39,8 +101,30 @@ export default function hasSameValues<T>(
     return false;
   }
 
-  const countsByValueA = countValues(iterableA);
-  const countsByValueB = countValues(iterableB);
+  const countsByValueA = countValuesSync(iterableA);
+  const countsByValueB = countValuesSync(iterableB);
+
+  if (countsByValueA.size !== countsByValueB.size) {
+    return false;
+  }
+
+  for (const [value, expectedCount] of countsByValueA) {
+    if (countsByValueB.get(value) !== expectedCount) {
+      return false;
+    }
+  }
+
+  return true;
+}
+
+async function hasSameValuesAsync<T>(
+  iterableA: Iterable<T> | AsyncIterable<T>,
+  iterableB: Iterable<T> | AsyncIterable<T>,
+): Promise<boolean> {
+  const [countsByValueA, countsByValueB] = await Promise.all([
+    countValuesEither(iterableA),
+    countValuesEither(iterableB),
+  ]);
 
   if (countsByValueA.size !== countsByValueB.size) {
     return false;