+++ test/hasSameValues.test.ts
@@ -1,8 +1,8 @@
-import { assertEquals } from "assert";
+import { assert, assertEquals } from "assert";
 
-import { hasSameValues } from "../mod.ts";
+import { asyncify, emptyAsyncIterable, hasSameValues } from "../mod.ts";
 
-const customIterable = {
+const nineEightSeven = {
   *[Symbol.iterator]() {
     yield 9;
     yield 7;
@@ -10,25 +10,43 @@ const customIterable = {
   },
 };
 
-Deno.test("returns true if the iterables have the same values", () => {
-  assertEquals(hasSameValues([], []), true);
-  assertEquals(hasSameValues([9, 8, 7], [7, 8, 9]), true);
-  assertEquals(hasSameValues([9, 9, 9], [9, 9, 9]), true);
+Deno.test("returns true if the iterables have the same values", async () => {
+  assert(hasSameValues([], []));
+  assert(hasSameValues([], new Set()));
+  assert(await hasSameValues(new Set(), emptyAsyncIterable));
+  assert(await hasSameValues(emptyAsyncIterable, emptyAsyncIterable));
 
-  assertEquals(hasSameValues(new Set([9, 8, 7]), [7, 8, 9]), true);
-  assertEquals(hasSameValues([9, 8, 7], new Set([7, 8, 9])), true);
-
-  assertEquals(hasSameValues([9, 8, 7], customIterable), true);
+  const nineEightSevens: Array<Iterable<number>> = [
+    [7, 8, 9],
+    [9, 8, 7],
+    new Set([9, 8, 7]),
+    nineEightSeven,
+  ];
+  for (const a of nineEightSevens) {
+    for (const b of nineEightSevens) {
+      assert(hasSameValues(a, b));
+      assert(hasSameValues(b, a));
+      assert(await hasSameValues(asyncify(a), b));
+      assert(await hasSameValues(asyncify(b), a));
+      assert(await hasSameValues(asyncify(a), asyncify(b)));
+    }
+  }
 });
 
 Deno.test(
   "returns false if the iterables don't contain the same values",
-  () => {
+  async () => {
     assertEquals(hasSameValues([1], []), false);
     assertEquals(hasSameValues([], [1]), false);
+    assertEquals(await hasSameValues(emptyAsyncIterable, [1]), false);
 
     assertEquals(hasSameValues([1, 2, 3], [2, 3]), false);
     assertEquals(hasSameValues([2, 3], [1, 2, 3]), false);
+    assertEquals(await hasSameValues([1, 2, 3], asyncify([2, 3])), false);
+    assertEquals(
+      await hasSameValues(asyncify([1, 2, 3]), asyncify([2, 3])),
+      false,
+    );
 
     assertEquals(hasSameValues([1, 2, 3], [1, 1, 2, 3]), false);
     assertEquals(hasSameValues([1, 1, 2, 3], [1, 2, 3]), false);
@@ -38,7 +56,7 @@ Deno.test(
 
     assertEquals(hasSameValues(new Set([1, 2, 3]), [1, 1, 2, 2, 3]), false);
 
-    assertEquals(hasSameValues(customIterable, [7, 8]), false);
-    assertEquals(hasSameValues([7, 8], customIterable), false);
+    assertEquals(hasSameValues(nineEightSeven, [7, 8]), false);
+    assertEquals(hasSameValues([7, 8], nineEightSeven), false);
   },
 );